refactor - 약속 상세 보기 화면 패키지명 변경
