+++ presentation/src/main/java/com/yapp/growth/presentation/ui/main/PlanzScreen.kt
@@ -41,7 +41,7 @@ import com.yapp.growth.presentation.ui.main.create.theme.ThemeScreen
 import com.yapp.growth.presentation.ui.main.create.title.TitleScreen
 import com.yapp.growth.presentation.ui.main.home.HomeScreen
 import com.yapp.growth.presentation.ui.main.manageplan.ManagePlanScreen
-import com.yapp.growth.presentation.ui.main.plandetail.DetailPlanScreen
+import com.yapp.growth.presentation.ui.main.detailPlan.DetailPlanScreen
 import com.yapp.growth.presentation.ui.main.sample.SampleScreen
 
 @Composable
